06_Projects/mac_compute_node/requirements.txt
aiohttp>=3.9.0
fastapi>=0.110.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
psutil>=5.9.8
requests>=2.31.0
numpy>=1.26.0
docker>=7.0.0
solana>=0.30.0
solders>=0.18.0
spl-token>=0.3.0
websockets>=12.0
watchdog>=3.0.0
httpx>=0.27.0
pydantic>=2.5.0
rich>=13.7.0
click>=8.1.0
web3>=6.15.0
